refactor: add unit tests for LocalRecentSearchDataSourceImpl
